+++ libs/common/src/auth/authorization/consts/metadata-keys.ts
@@ -1,0 +1,9 @@
+/**
+ * @private
+ */
+export const REQUIRED_PERMISSIONS = 'AUTH_REQUIRED_PERMISSIONS';
+
+/**
+ * @private
+ */
+export const IS_PUBLIC = 'AUTH_IS_PUBLIC';