+++ libs/common/src/auth/authorization/public-api.ts
@@ -1,0 +1,11 @@
+export { AUTH_USER_RESOLVER } from './consts/injection-tokens';
+
+export * from './decorators/public.decorator';
+
+export * from './guards/permissions.guard';
+
+export * from './services/public-access-advisor.service';
+
+export * from './types/auth-user-resolver';
+export * from './types/authorization.options';
+export * from './types/authorization.service';