+++ libs/common/src/auth/utility/types/password-utility.options.ts
@@ -1,0 +1,8 @@
+export interface PasswordUtilityOptions {
+
+  /**
+   * Passed to `bcrypt.hash`, default is 10
+   */
+  saltOrRounds?: number;
+}
+