+++ libs/common/src/openapi/index.ts
@@ -2,6 +2,7 @@ export * from './decorators/api-enum';
 export * from './decorators/api-self-tag';
 export * from './decorators/query-result.response';
 
+export * from './factories/require-authorization.factory';
 export * from './factories/swagger-ui.factory';
 export * from './factories/validate.factory';