+++ libs/common/changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Interface and class for CRUD
 - More OpenAPI utils, CRUD and validation related
+- Authorization and password utilities
 
 ## [0.1.0] - 2024-02-04